Split Web App and Database configuration validation
